+++ extensions/github/src/my-pull-requests-menu.tsx
@@ -25,18 +25,18 @@ function getMaxPullRequestsPreference(): number {
   return getBoundedPreferenceNumber({ name: "maxitems" });
 }
 
-function getPullRequestStatusIcon(pr: PullRequestFieldsFragment): { source: Icon | string } {
+function getPullRequestStatusIcon(pr: PullRequestFieldsFragment): Icon | string {
   const pullRequestStatus = pr.commits.nodes ? pr.commits.nodes[0]?.commit.statusCheckRollup?.state : null;
   switch (pullRequestStatus) {
     case "SUCCESS":
-      return { source: Icon.Check };
+      return Icon.Check;
     case "ERROR":
     case "FAILURE":
-      return { source: Icon.Xmark };
+      return Icon.Xmark;
     case "PENDING":
-      return { source: Icon.Clock };
+      return Icon.Clock;
     default:
-      return { source: "pull-request-open.svg" };
+      return "pull-request-open.svg";
   }
 }
 
@@ -114,7 +114,7 @@ function MyPullRequestsMenu() {
                 <MenuBarItem
                   key={i.id}
                   title={`#${i.number} ${i.title}`}
-                  icon={getPullRequestStatusIcon(i)}
+                  icon={{ source: getPullRequestStatusIcon(i), tintColor: Color.PrimaryText }}
                   tooltip={i.repository.nameWithOwner}
                   onAction={() => open(i.permalink)}
                 />