+++ extensions/github/CHANGELOG.md
@@ -1,9 +1,10 @@
 # GitHub Changelog
 
-## [Fix icons for Notifications and Unread Notifications commands] - 2024-04-05
+## [Fix various icons and colors] - 2024-04-05
 
 - Fixes icons for issues and pull requests in `Notifications` and `Unread Notifications` commands.
 - Standardizes the color for all the accessories in `My Pull Requests`.
+- Fixes an issue where the pull request icon color for the `My Pull Requests Menu Bar` was not adjusted to the dark mode.
 
 ## [Fix icon] - 2024-04-02